Move src files under protocols/aave-v3/src
